rename generated GRPC files to avoid false detection by pytest
